Improved existing buttons and added more
The easy, medium, and hard difficulty buttons now lead to a screen containing the reset, restart, and exit buttons. Restart and exit are fully functional. When you hover over the buttons, the center rectangle's color changes to notify the user that the button is about to be pressed. I made sure one button's functioning wouldn't get in the way of the other buttons.