Logic for def valid_in_box~
checks if it's valid to place a given number (num) in a specific box of the 2D board(does this by iterating through each cell using a nested loop). It examines each cell in the box and returns False if the number is already present in that box, indicating an invalid move. If no matching number is found, it implies the move is valid, and the method returns None